add permission checks for nir model functions

+++ src/features/nir-models/actions/nir-model.ts
@@ -2,6 +2,8 @@
 
 import { z } from "zod";
 
+import { getCurrentUser } from "@/lib/currentUser";
+import { hasPermission } from "@/permissions/general";
 import {
   deleteNirModel as deleteNirModelDb,
   insertNirModel,
@@ -13,7 +15,10 @@ export async function createNirModel(
 ) {
   const { success, data } = nirModelSchema.safeParse(unsafeData);
 
-  if (!success) {
+  const user = await getCurrentUser();
+  const canCreateNirModel = hasPermission(user?.role, "createNirModel");
+
+  if (!success || !canCreateNirModel) {
     return {
       error: true,
       message: "There was an error creating the NIR model",
@@ -33,6 +38,15 @@ export async function createNirModel(
 
 export async function deleteNirModel(id: number) {
   try {
+    const user = await getCurrentUser();
+    const canDeleteNirModel = hasPermission(user?.role, "deleteNirModel");
+
+    if (!canDeleteNirModel) {
+      return {
+        error: true,
+        message: "You do not have permission to delete NIR models.",
+      };
+    }
     await deleteNirModelDb({ id });
     return { error: false, message: "Successfully deleted the NIR model" };
   } catch (error) {
